Cleanup workspace fetching
